requirements.txt
asgiref==3.2.3
bleach==1.4.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.11
django-connection-url==0.1.2
django-filter==2.2.0
django-guardian==2.2.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==19.6.0
html5lib==0.9999999
httpie==2.0.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.6.1
python-decouple==3.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
Unipath==1.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
bs4